Change onWorksheetRemoved currentWorkbook check
There is no workbook on the removed worksheet. The Excel API will either need to change to include workbook when it creates this.objectInstance or the event can be changed to include the workbook alongside the removed worksheet. The check has been updated to use the target which is the workbook until the api changes to include workbook in the event (or removed worksheet).