Updated to clear rtd cell on provider close
Updated dependencies for the add in (around packaging/zipping) to be latest stable.
Removed the option to not clear rtd topic cells when calling dispose as we now clear the cell value when the provider is shutdown.
Updated output of index.d.ts in src so it could be used for typing information.
Updated documentation.